% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{cal.eigen.fit}
\alias{cal.eigen.fit}
\title{(Internal function) Calculae a vector of EigenFit values, internally used for
parallelization}
\usage{
cal.eigen.fit(eigen.value)
}
\arguments{
\item{eigen.value}{A vector of Eigenvalues return from \code{\link{svd}}
($d), \code{rARPACK::svds} ($d),  \code{\link{eigen}}
($values) or \code{rARPACK::eigs} ($values).}
}
\value{
A vector of all possible values for EigenFit.
}
\description{
(Internal function) Calculae a vector of EigenFit values, internally used for
parallelization
}
