
\name{IPCAPS-package}
\alias{IPCAPS-package}
\alias{IPCAPS}
\docType{package}
\title{
  IPCAPS - Iterative Pruning to CApture Population Structure
}
\description{
  Determining fine population structure using iterative pruning.
}
\details{
\tabular{ll}{
Package: \tab IPCAPS\cr
Type: \tab Package\cr
Version: \tab 0.43.0\cr
Depends: \tab R (>= 3.0.0), Matrix, expm, fpc, Rmixmod, LPCM, apcluster, rARPACK\cr
Date: \tab 2017-07-21\cr
License: \tab GPL 3\cr
}
This package containts a function \emph{ipcaps} for unsuppervised clustering.
}
\author{
Kridsadakorn Chaichoompu, Fentaw Abegaz Yazew, Sissades Tongsima, Philip James Shaw, Anavaj Sakuntabhai, Luisa Pereira, and Kristel Van Steen

Maintainer: Kridsadakorn Chaichoompu \emph{<kridsadakorn.chaichoompu@ulg.ac.be>}
}
\references{
  
}
\keyword{ clustering}
\keyword{ population }
\keyword{ iterative pruning }
\keyword{ ipPCA }
\keyword{ PCA }
\examples{

}
