% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pasre.categorical.data}
\alias{pasre.categorical.data}
\title{(Internal function) Manipulate categorical input files}
\usage{
pasre.categorical.data(files)
}
\arguments{
\item{files}{IPCAPS supports categorical data, which rows represent features
and columns represent subjects or individuals. Each column needs to be
separated by a space or a tab. A big text file should be divided into smaller
files to load faster. For instance, to input 3 files, use as: files=c(
'input1.txt', 'input2.txt', 'input3.txt').}
}
\value{
A data frame of input data
}
\description{
(Internal function) Manipulate categorical input files
}
