% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC}
\alias{PC}
\title{Synthetic dataset containing the top 10 principal components (PC) from the
dataset \code{raw.data}}
\format{
A matrix with 10 columns and 1,004 rows
}
\usage{
data(ipcaps_example)
}
\description{
A dataset contains a numeric matrix of 1,004 rows and 10 columns of top 10
PCs calculated from the dataset \code{raw.data}. The PCs were calculated
using linear principal component analysis (PCA), see more datails at
\code{KRIS::cal.pc.linear}
}
\seealso{
\code{\link{raw.data}} and \code{\link{label}}
}
\keyword{PC}
