% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.each.node.R
\name{process.each.node}
\alias{process.each.node}
\title{(Internal function) Perform the iterative process for each node}
\usage{
process.each.node(node, work.dir)
}
\arguments{
\item{node}{An integer representing the current node number which is being
analyzed.}

\item{work.dir}{A working directory.}
}
\value{
\code{NULL}
}
\description{
(Internal function) Perform the iterative process for each node
}
