% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{replace.missing}
\alias{replace.missing}
\title{(Internal function) Replace missing values by specified values, internally
used for parallelization}
\usage{
replace.missing(X, missing = NA, rep)
}
\arguments{
\item{X}{A vector of data}

\item{missing}{The old characters representing a missing value, for example, 'NA' or '-'}

\item{rep}{The new characters to replace the old characters, for example, 'NULL' or '-1'}
}
\value{
A vector of data with replaced character for miss values
}
\description{
(Internal function) Replace missing values by specified values, internally
used for parallelization
}
