% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.discriminator.R
\name{top.discriminator}
\alias{top.discriminator}
\title{Detecting top discriminators between two groups}
\usage{
top.discriminator(cluster.obj, group1, group2, bim.file,
  use.node.number = FALSE, num.top = 100)
}
\arguments{
\item{cluster.obj}{The object which is returned from \code{\link{ipcaps}}.}

\item{group1}{To specify the first group number to be compared. (also see
\code{use.node.number})}

\item{group2}{To specify the second group number to be compared. (also see
\code{use.node.number})}

\item{bim.file}{Option: In case that SNP information is not provided to
\code{\link{ipcaps}}, an absolute path of SNP information file is required.
It needs to be in PLINK format (bim). See more details at:
\url{http://zzz.bwh.harvard.edu/plink/data.shtml}.}

\item{use.node.number}{To specify whether a group number or a node number
is be used. If TRUE, a node nubmer is used instead. Default = FALSE.}

\item{num.top}{A number of top Fst SNPs to be returned. Default = 100.}
}
\value{
The returned value is a data.frame of SNP information sorting by Fst
in descending order, which contains 7 columns, chr, SNP, centimorgans,
position, allele1, allele2, and Fst. The column 1-6 are SNP information from
the bim file. The column Fst contains estimated Fst between group1 and group2.
}
\description{
Detects top discriminators that contribute to
group separation based on the fixation index (Fst).
}
