% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{cal.eigen.fit}
\alias{cal.eigen.fit}
\title{(Internal function) For parallelization purpose, to calculae a vector of EigenFit values}
\usage{
cal.eigen.fit(eigen.value)
}
\arguments{
\item{eigen.value}{A vector of Eigenvalues return from \code{\link{svd}}
($d), \code{\link{rARPACK::svds}} ($d),  \code{\link{eigen}}
($values) or \code{\link{rARPACK::eigs}} ($values).}
}
\value{
A vector of all possible values for EigenFit.
}
\description{
(Internal function) For parallelization purpose, to calculae a vector of EigenFit values
}
\examples{
 X <- sort(runif(10, min = 0, max = 3), decreasing = TRUE)
 print(cal.eigen.fit(X))
}
