% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{diff.eigen.fit}
\alias{diff.eigen.fit}
\title{(Internal function) Calculate a vector of different values from a vector of
EigenFit values, internally used for parallelization}
\usage{
\method{diff}{eigen.fit}(eigen.value)
}
\arguments{
\item{eigen.value}{A vector of Eigenvalues return from \code{svd} ($d),
\code{rARPACK::svds} ($d), \code{eigen} ($values) or \code{rARPACK::eigs} ($values).}
}
\value{
A vector of different values from a vector of EigenFit values
}
\description{
(Internal function) Calculate a vector of different values from a vector of
EigenFit values, internally used for parallelization
}
