% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{diff.eigen.fit}
\alias{diff.eigen.fit}
\title{(Internal function) Calculate a vector of different values from a vector of
EigenFit values, internally used for parallelization}
\usage{
\method{diff}{eigen.fit}(eigen.value)
}
\arguments{
\item{eigen.value}{A vector of Eigenvalues return from \code{\link{svd}}
($d), \code{\link{rARPACK::svds}} ($d),  \code{\link{eigen}}
($values) or \code{\link{rARPACK::eigs}} ($values).}
}
\value{
A vector of different values from a vector of EigenFit values
}
\description{
(Internal function) Calculate a vector of different values from a vector of
EigenFit values, internally used for parallelization
}
\examples{
 X <- sort(runif(10, min = 0, max = 3), decreasing = TRUE)
 print(diff.eigen.fit(X))
}
