% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{diff.xy}
\alias{diff.xy}
\title{(Internal function) Check the different value of X and Y, internally used for
parallelization}
\usage{
\method{diff}{xy}(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
The different number of \code{x} and \code{y}.
}
\description{
(Internal function) Check the different value of X and Y, internally used for
parallelization
}
\examples{
print(diff.xy(2.4,6.7))
print(diff.xy(8.23,3.5))
print(diff.xy(4.3,4.3))
}
