% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{replace.missing}
\alias{replace.missing}
\title{(Internal function) Replace missing values by specified values, internally
used for parallelization}
\usage{
replace.missing(X, missing = NA, rep)
}
\arguments{
\item{X}{A vector of data}

\item{missing}{The old characters representing a missing value, for example, 'NA' or '-'}

\item{rep}{The new characters to replace the old characters, for example, 'NULL' or '-1'}
}
\value{
A vector of data with replaced character for miss values
}
\description{
(Internal function) Replace missing values by specified values, internally
used for parallelization
}
\examples{
X <- runif(20, min = 0, max = 3)
X[c(2,4,8,13,17)] <- NA
missing <- NA
replace.char <- rep(0,20)
res <- replace.missing(X, missing.char, replace.char)
print(X)
print(res)
}
